I fixed problem in model.py
I change data type of values column to FloatField from IntegerFiled becuase it could not hold whole float numbers.

+++ csv_processor_core/models.py
@@ -4,7 +4,7 @@ class Farms(models.Model):
     location = models.CharField(max_length=150)
     date_time = models.DateTimeField()
     sensor_type = models.CharField(max_length=50)
-    values = models.IntegerField()
+    values = models.FloatField()
 
 
     def __str__(self):
